Development

+++ src/main/java/net/bplaced/clayn/cfs/impl/local/CFSSimpleFileImpl.java
@@ -49,6 +49,10 @@ public class CFSSimpleFileImpl implements SimpleFile
     @Override
     public void delete() throws IOException
     {
+        if (!exists())
+        {
+            return;
+        }
         Files.delete(realFile);
     }
 
@@ -136,8 +140,7 @@ public class CFSSimpleFileImpl implements SimpleFile
     @Override
     public String toString()
     {
-        return parent.toString()+getName();
+        return parent.toString() + getName();
     }
-    
-    
+
 }
